backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-text-splitters
langchain-community
langchain-huggingface
langchain-google-genai
python-dotenv
supabase
PyMuPDF
sentence-transformers
faiss-cpu